# docs/swagger.yaml
definitions:
...
  api_response.CreateTicketResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      ticket:
        $ref: '#/definitions/models.Ticket'
    type: object
...
  api_response.DeleteTicketResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  api_response.GetTicketResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      ticket:
        $ref: '#/definitions/models.Ticket'
    type: object
  api_response.GetTicketsResponse:
    properties:
      message:
        type: string
      pagination:
        $ref: '#/definitions/types.Pagination'
      status:
        type: integer
      success:
        type: boolean
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
    type: object
...
  api_response.UpdateTicketResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      ticket:
        $ref: '#/definitions/models.Ticket'
    type: object
...
  models.AssignTicketRequest:
    properties:
      assignee_id:
        type: integer
    required:
    - assignee_id
    type: object
  models.Comment:
    properties:
      comment:
        maxLength: 1000
        type: string
      created_at:
        type: string
      id:
        type: integer
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticket_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    required:
    - comment
    type: object
...
  models.CreateTicketRequest:
    properties:
      assigned_by_id:
        type: integer
      assignee_id:
        type: integer
      description:
        maxLength: 1000
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      status:
        enum:
        - open
        - in_progress
        - closed
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  models.Ticket:
    properties:
      assigned_by:
        $ref: '#/definitions/models.User'
      assigned_by_id:
        description: Who assigned the ticket
        type: integer
      assignee:
        $ref: '#/definitions/models.User'
      assignee_id:
        type: integer
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      created_by_id:
        description: Who created the ticket
        type: integer
      createdBy:
        $ref: '#/definitions/models.User'
      description:
        maxLength: 1000
        type: string
      id:
        type: integer
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      status:
        enum:
        - open
        - in_progress
        - closed
        type: string
      title:
        maxLength: 255
        minLength: 5
        type: string
      updated_at:
        type: string
    required:
    - title
    type: object
...
  models.UpdateTicketRequest:
    properties:
      assigned_by_id:
        type: integer
      assignee_id:
        type: integer
      description:
        maxLength: 1000
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      status:
        enum:
        - open
        - in_progress
        - closed
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  types.Pagination:
    properties:
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: API documentation
  title: Example API
  version: "1.0"
paths:
...
  /api/v1/tickets:
    get:
      description: Retrieve tickets with optional pagination (page, size) and filter
        by status, priority, assignee_id, reporter_id
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: size
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by priority
        in: query
        name: priority
        type: string
      - description: Filter by assignee_id
        in: query
        name: assignee_id
        type: integer
      - description: Filter by reporter_id
        in: query
        name: reporter_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.GetTicketsResponse'
      security:
      - BearerAuth: []
      summary: Get tickets with pagination and filters
      tags:
      - tickets
    post:
      consumes:
      - application/json
      description: Create a new ticket item
      parameters:
      - description: Ticket object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateTicketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.CreateTicketResponse'
        "400":
          description: Invalid JSON
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new ticket
      tags:
      - tickets
  /api/v1/tickets/{id}:
    delete:
      description: Delete a specific ticket by its ID
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.DeleteTicketResponse'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Ticket not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete ticket by ID
      tags:
      - tickets
    get:
      description: Fetch a specific ticket by its ID
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.GetTicketResponse'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Ticket not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get ticket by ID
      tags:
      - tickets
    patch:
      consumes:
      - application/json
      description: Update a ticket item by ID
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ticket object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AssignTicketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.UpdateTicketResponse'
        "400":
          description: Invalid JSON
          schema:
            type: string
        "404":
          description: Ticket not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update an existing ticket
      tags:
      - tickets
  /api/v1/tickets/{id}/assign:
    patch:
      consumes:
      - application/json
      description: Assign a ticket item by ID
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ticket object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTicketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.UpdateTicketResponse'
        "400":
          description: Invalid JSON
          schema:
            type: string
        "404":
          description: Ticket not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Assign an existing ticket
      tags:
      - tickets
  /api/v1/tickets/{id}/comments:
    get:
      description: Fetch a specific ticket comments by its ID
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.GetTicketResponse'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Ticket not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get ticket comments by ID
      tags:
      - tickets
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"